feat(stream): move stories to the dedicated storybook workspace
