+++ storybook/package.json
@@ -33,6 +33,7 @@
     "@nivo/radial-bar": "workspace:*",
     "@nivo/sankey": "workspace:*",
     "@nivo/scatterplot": "workspace:*",
+    "@nivo/stream": "workspace:*",
     "@nivo/tooltip": "workspace:*",
     "@nivo/waffle": "workspace:*",
     "@babel/preset-env": "^7.21.5",
